End-to-end with exception of Acumatica not creating one-to-many Transactions

+++ Push.Shopify/Api/Payout/PayoutDetail.cs
@@ -12,9 +12,9 @@ namespace Push.Shopify.Api.Payout
         public long? payout_id { get; set; }
         public string payout_status { get; set; }
         public string currency { get; set; }
-        public decimal amount { get; set; }
-        public decimal fee { get; set; }
-        public decimal net { get; set; }
+        public double amount { get; set; }
+        public double fee { get; set; }
+        public double net { get; set; }
         public long source_id { get; set; }
         public string source_type { get; set; }
         public long? source_order_id { get; set; }
